utils: separate concerns of github and openQA APIs
Separate the concerns of github stuff and openQA stuff into each
file so later it's easier to integrate the openqa_investigator.py
script.
