+++ utils/lib/common.py
@@ -1,0 +1,8 @@
+LABEL_OK = 'openqa-ok'
+LABEL_FAILED = 'openqa-failed'
+COMMENT_TITLE = "# OpenQA test summary"
+
+GITHUB_BASE_PREFIX = "https://github.com"
+
+ISSUE_TITLE_PREFIX = "OpenQA test result for build "
+
